Add some simple tests and snapshots
