+++ package.json
@@ -11,7 +11,8 @@
     "test:ci:all": "lerna run test:ci",
     "test:all": "lerna run test",
     "tsc:all": "lerna run tsc:all",
-    "prettier": "lerna run prettier"
+    "prettier": "lerna run prettier",
+    "postinstall": "lerna bootstrap"
   },
   "repository": {
     "type": "git",
