--- packages/snapshot/__tests__/default.spec.ts
@@ -1,1 +1,0 @@
-test("empty", () => {});
