+++ packages/snapshot/lib/index.d.ts
@@ -1,2 +1,5 @@
 export * from "./parser";
+export * from "./test/Snapshot";
+export * from "./test/SnapshotDiff";
+export * from "./test/unparse";
 //# sourceMappingURL=index.d.ts.map
