+++ packages/snapshot/package.json
@@ -30,7 +30,10 @@
     "nearley": "^2.19.1"
   },
   "devDependencies": {
+    "@types/jest": "^25.1.3",
     "@types/nearley": "^2.11.1",
-    "prettier": "^1.19.1"
+		"jest": "^25.1.0",
+    "prettier": "^1.19.1",
+		"ts-jest": "^25.2.1"
   }
 }
