+++ packages/snapshot/src/index.ts
@@ -1,1 +1,4 @@
 export * from "./parser";
+export * from "./test/Snapshot";
+export * from "./test/SnapshotDiff";
+export * from "./test/unparse";
